--- InterfaceQt/resource.qrc
@@ -1,7 +1,0 @@
-<RCC>
-    <qresource prefix="/rec">
-        <file>img/zrk.png</file>
-        <file>img/plane_115212.png</file>
-        <file>img/gyrocopter.png</file>
-    </qresource>
-</RCC>
